# Prevent loader to add default on frozen objects