If no default is defined on exports, loader.js will try to add one.
This can cause troubles with some bundles exported from rollupjs.