@kratiahuja 2016-10-21
@apezel isn't this the same issue as this: https://github.com/ember-cli/loader.js/issues/91