@stefanpenner 2016-10-21
whats the performance impact of this? It should be quick, but I don't trust my intuition.