@stefanpenner 2016-10-21
To accept this, it also requires test coverage