@apezel 2016-10-21
@stefanpenner Tests are ok.

Running the benchmark tool on examples/ember.js i didn't get significant differences : 

Without frozen detection : 

Benchmark
  file: benchmarks/scenarios/ember.js
  total:  2224ms
  per op:  222.4ms

With Object.isFrozen : 

Benchmark
  file: benchmarks/scenarios/ember.js
  total:  2233ms
  per op:  223.3ms