@rwjblue 2017-03-05
Added tests for frozen objects and tweaked check to handle all non-extensible objects in https://github.com/ember-cli/loader.js/pull/113.